feat: migrate AppRoutes.spec
